+++ common/src/states/spin_melee.rs
@@ -118,7 +118,7 @@ impl CharacterBehavior for Data {
                         timer: tick_attack_or_default(data, self.timer, None),
                         ..*self
                     });
-                } else if update.energy.current() as f32 >= self.static_data.energy_cost
+                } else if update.energy.current() >= self.static_data.energy_cost
                     && (self.consecutive_spins < self.static_data.num_spins
                         || (self.static_data.is_infinite
                             && self.static_data.ability_info.input.map_or(false, |input| input_is_pressed(data, input))))
