+++ voxygen/anim/src/biped_large/shockwave.rs
@@ -43,7 +43,7 @@ impl Animation for ShockwaveAnimation {
         let mut next = (*skeleton).clone();
 
         let (move1, move1pow, move2, move3) = match stage_section {
-            Some(StageSection::Buildup) => (anim_time, anim_time.powf(0.25) as f32, 0.0, 0.0),
+            Some(StageSection::Buildup) => (anim_time, anim_time.powf(0.25), 0.0, 0.0),
             Some(StageSection::Action) => (1.0, 1.0, anim_time, 0.0),
             Some(StageSection::Recover) => (1.0, 1.0, 1.0, anim_time),
             _ => (0.0, 0.0, 0.0, 0.0),
