+++ world/src/site/settlement/mod.rs
@@ -232,8 +232,8 @@ impl Settlement {
         let river_offs = Vec2::new(rng.gen_range(-3..4), rng.gen_range(-3..4));
 
         for x in (0..100).map(|e| e as f32 / 100.0) {
-            let theta0 = x as f32 * f32::consts::PI * 2.0;
-            let theta1 = (x + 0.01) as f32 * f32::consts::PI * 2.0;
+            let theta0 = x * f32::consts::PI * 2.0;
+            let theta1 = (x + 0.01) * f32::consts::PI * 2.0;
 
             let pos0 = (river_dir * radius + Vec2::new(theta0.sin(), theta0.cos()) * radius)
                 .map(|e| e.floor() as i32)
